requirements.txt
oemof-solph
matplotlib
seaborn
pandas
ipykernel
openpyxl